Improve module augmentation for `language` plugin
