+++ packages/ckeditor5-language/src/textpartlanguageediting.ts
@@ -13,8 +13,6 @@ import { Plugin, type Editor } from 'ckeditor5/src/core';
 import TextPartLanguageCommand from './textpartlanguagecommand';
 import { stringifyLanguageAttribute, parseLanguageAttribute } from './utils';
 
-import './textpartlanguageconfig';
-
 /**
  * The text part language editing.
  *
